Tests/LibWeb/Text/expected/UIEvents/KeyEvent-tab-cancelled-does-not-propagate.txt

input1 keydown Tab
cancel keydown Tab
<INPUT id="input1">
input1 keydown Tab
<INPUT id="input2">
